core: changed file name, added emoji/description, added switch feature for single roles
